requirements.txt
djitellopy
opencv-python
Pillow
tk
